flask-demo/requirements.txt
Flask==0.12.2
tensorflow==1.11.0
numpy==1.15.3
pillow==5.4.1
